+++ packages/core/test/SchemaField.test.jsx
@@ -52,7 +52,7 @@ describe('SchemaField', () => {
         schemaUtils,
         translateString: englishStringTranslator,
         globalUiOptions: undefined,
-        experimental_componentUpdateStrategy: 'customDeep',
+        globalFormOptions: undefined,
       });
     });
     it('should provide expected registry with globalUiOptions as prop', () => {
@@ -87,7 +87,7 @@ describe('SchemaField', () => {
         schemaUtils,
         translateString: englishStringTranslator,
         globalUiOptions: { copyable: true },
-        experimental_componentUpdateStrategy: 'customDeep',
+        globalFormOptions: undefined,
       });
     });
   });